requirements.txt
autopep8==1.3.3
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.8.2
env==0.1.0
flake8==3.5.0
idna==2.6
isort==4.2.15
lxml==4.1.1
Markdown==2.6.11
mccabe==0.6.1
pipenv==2018.5.18
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.3
requests==2.18.4
selenium==3.8.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
